Add files via upload
==ship scout==
This unit pretty much does its roll, but i removed much of its sonar power (800 los) -> 200, to encourage submarine attacks.
Damage 34 -> 15.
Added Timeslow damage factor x2
The unit changes for this are not substantial since its roll is mainly used as just a scout. Its disarm/slow abilities just keep it a little bit more usefull later in the game since spamming light units to spot "invisible" doesn't happen on the ocean.
--Cutter--(scout)
Cheap, fast, and fragile, this Picket Ship is a good scout usefull for spotting units on the surface of the ocean. Though it has a weak attack in terms of damage, it does have some slowing and disarming potential.
--------------------


==Ship Raider.==
I felt that a raider should not be a sub counter. Since that means it will have not only torpedos, but also sonar, speed and manuevarbility. This can see someone trying to build some stealthy "underwater" subs, and they are getting killed by standard raiders. This also puts amp